+++ src/test/java/com/mobble/mobbleserver/infrastructure/persistence/meeting/MeetingRepositoryTest.java
@@ -2,7 +2,7 @@ package com.mobble.mobbleserver.infrastructure.persistence.meeting;
 
 import com.mobble.mobbleserver.config.QueryDslConfig;
 import com.mobble.mobbleserver.domain.ClubCategory.ClubCategory;
-import com.mobble.mobbleserver.domain.club.core.Club;
+import com.mobble.mobbleserver.domain.club.Club;
 import com.mobble.mobbleserver.domain.clubMember.ClubMember;
 import com.mobble.mobbleserver.domain.clubMember.ClubMemberRole;
 import com.mobble.mobbleserver.domain.clubMember.JoinStatus;
