+++ src/test/java/com/mobble/mobbleserver/support/fixture/clubMember/ClubMemberTestFixture.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.support.fixture.clubMember;
 
-import com.mobble.mobbleserver.domain.club.core.Club;
+import com.mobble.mobbleserver.domain.club.Club;
 import com.mobble.mobbleserver.domain.clubMember.ClubMember;
 import com.mobble.mobbleserver.domain.clubMember.ClubMemberRole;
 import com.mobble.mobbleserver.domain.clubMember.JoinStatus;
